+++ migration-library/src/main/kotlin/com/quadient/migration/api/dto/migrationmodel/Image.kt
@@ -4,7 +4,7 @@ import com.quadient.migration.shared.ImageOptions
 import com.quadient.migration.shared.ImageType
 import com.quadient.migration.shared.MetadataPrimitive
 import com.quadient.migration.shared.SkipOptions
-import kotlinx.datetime.Instant
+import kotlin.time.Instant
 
 data class Image @JvmOverloads constructor(
     override val id: String,