+++ migration-library/src/main/kotlin/com/quadient/migration/service/deploy/DeploymentResult.kt
@@ -5,7 +5,7 @@ package com.quadient.migration.service.deploy
 import com.quadient.migration.api.InspireOutput
 import com.quadient.migration.api.repository.StatusTrackingRepository
 import com.quadient.migration.shared.DocumentObjectType
-import kotlinx.datetime.Instant
+import kotlin.time.Instant
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid