+++ migration-library/src/main/kotlin/com/quadient/migration/service/deploy/DesignerDeployClient.kt
@@ -15,10 +15,10 @@ import com.quadient.migration.service.inspirebuilder.DesignerDocumentObjectBuild
 import com.quadient.migration.service.ipsclient.IpsService
 import com.quadient.migration.service.ipsclient.OperationResult
 import com.quadient.migration.shared.DocumentObjectType
-import kotlinx.datetime.Clock
-import org.jetbrains.exposed.v1.core.SqlExpressionBuilder.eq
-import org.jetbrains.exposed.v1.core.SqlExpressionBuilder.inList
+import kotlin.time.Clock
 import org.jetbrains.exposed.v1.core.and
+import org.jetbrains.exposed.v1.core.eq
+import org.jetbrains.exposed.v1.core.inList
 import org.jetbrains.exposed.v1.json.extract
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
@@ -51,8 +51,8 @@ class DesignerDeployClient(
     }
 
     override fun deployDocumentObjectsInternal(documentObjects: List<DocumentObject>): DeploymentResult {
-        val deploymentId = kotlin.uuid.Uuid.random()
-        val deploymentTimestamp = kotlinx.datetime.Clock.System.now()
+        val deploymentId = Uuid.random()
+        val deploymentTimestamp = Clock.System.now()
         val deploymentResult = DeploymentResult(deploymentId)
 
         val orderedDocumentObject = deployOrder(documentObjects)