+++ migration-library/src/main/kotlin/com/quadient/migration/service/deploy/ProgressReport.kt
@@ -5,7 +5,7 @@ package com.quadient.migration.service.deploy
 import com.quadient.migration.api.dto.migrationmodel.DocumentObject
 import com.quadient.migration.api.dto.migrationmodel.Image
 import com.quadient.migration.api.dto.migrationmodel.Attachment
-import kotlinx.datetime.Instant
+import kotlin.time.Instant
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid