+++ migration-library/src/main/kotlin/com/quadient/migration/shared/IcmMetadata.kt
@@ -1,7 +1,7 @@
 package com.quadient.migration.shared
 
-import kotlinx.datetime.Instant
-import kotlinx.datetime.toJavaInstant
+import kotlin.time.Instant
+import kotlin.time.toJavaInstant
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable